+++ src/layout/navBars/breadcrumb/index.vue
@@ -109,7 +109,6 @@ export default {
 	align-items: center;
 	padding-right: 15px;
 	background: var(--bg-topBar);
-	overflow: hidden;
 	border-bottom: 1px solid #f1f2f3;
 }
 </style>